Move deprecated tests to DeprecatedDatasetAggregatorSuite
